Update service to scan all running containers on service start
When a host restarts, there is no guarantee that this service will start before the ones that you want to have modified. This then leads to those started containers not having the required permissions to access the /dev nodes.
This update runs a quick parse over the running containers and for each of them with a /dev mount, it will apply the device rules.